requirements.txt
clamd>=1.0.2
yara-python>=4.3.1
vt-py>=0.18.0
pyudev>=0.24.1; sys_platform == "linux"
wmi>=1.5.1; sys_platform == "win32"
PyYAML>=6.0.1
pytest>=8.0.0
pytest-mock>=3.12.0
